refactor and preparation for pymchelper code
